tools/backup_system/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
jinja2>=3.1.3
requests>=2.31.0
